[ch]-seperate group validation from client validation file
